+++ src/engine/sprite.rs
@@ -1,12 +1,12 @@
 #[derive(Copy, Clone)]
 pub struct Sprite {
-    pub width: f32,
-    pub height: f32,
+    pub width: u32,
+    pub height: u32,
     pub tex_coords: [[f32; 2]; 4],
 }
 
 impl Sprite {
-    pub fn new(width: f32, height: f32, tex_coords: [[f32; 2]; 4]) -> Sprite {
+    pub fn new(width: u32, height: u32, tex_coords: [[f32; 2]; 4]) -> Sprite {
         Sprite {
             width: width,
             height: height,
